feat: 투명 textarea
