+++ src/stories/TextArea.stories.ts
@@ -1,48 +1,29 @@
-import CommomTextArea from "@/app/components/input/CommomTextArea";
-
-enum TextAreaSize {
-  SMALL = "w-[327px] h-[124px]",
-  LARGE = "w-[640px] h-[196px]",
-}
+import TransparentTextArea from "@/app/components/input/TransparentTextArea";
 
 enum BgColor {
-  WHITE = "bg-white",
-  TRANSPARENT = "bg-transparent",
+  WHITE = "bg-background-200",
+  TRANSPARENT = "bg-transparent border-[0.5px]",
 }
 
 enum BorderColor {
-  DEFAULT = "border-gray-200",
-  HOVER = "border-gray-300",
-  FOCUS = "border-primary-orange-300",
+  DEFAULT = "border-grayscale-200",
+  NONE = "border-0",
   ERROR = "border-state-error",
 }
 
-enum TextColor {
-  GRAY = "text-gray-400",
-  BLACK = "text-black-400",
-}
-
 export default {
-  title: "Test/TextArea",
-  component: CommomTextArea,
+  title: "Test/TransparentTextArea",
+  component: TransparentTextArea,
   args: {
     placeholder: "입력하세요",
     disabled: false,
-    borderWidth: "border-[0.5px]",
-    borderColor: BorderColor.DEFAULT,
-    textColor: TextColor.GRAY,
-    size: TextAreaSize.SMALL,
-    bgColor: BgColor.TRANSPARENT,
   },
 };
-
-export const textArea_default = {};
+export const defalut = {};
 
 export const textArea_error = {
   args: {
     borderColor: BorderColor.ERROR,
-    textColor: TextColor.BLACK,
-    errorAlign: "text-right",
     errorMessage: "가게 이름(상호명)을 필수로 입력해주세요", // 필수값 검증
   },
 };
